EcoduleAppをAuth部分とApp部分に切り分け
