+++ app/src/main/java/com/example/ecodule/ui/UserData.kt
@@ -6,5 +6,4 @@ import kotlinx.serialization.Serializable
 data class UserData(
     val id: String,
     val email: String,
-    val accessToken: String
 )
