--- app/src/main/java/com/example/ecodule/ui/account/model/UserData.kt
@@ -1,7 +1,0 @@
-package com.example.ecodule.ui.account.model
-
-data class UserData(
-    val id: String,
-    val email: String,
-    val accessToken: String
-)
